refactor: 履歴純関数を共有モジュールに移動 (#36)
